refactor: 디자인시스템에서 프리뷰를 보기 위한 원형 의존성 에러 해결
