+++ core/designsystem/build.gradle.kts
@@ -8,6 +8,7 @@ android {
 }
 
 dependencies {
+    implementation(projects.core.ui)
     implementation(projects.core.util.android)
     implementation(projects.core.util.jvm)
 }
