+++ core/designsystem/src/main/kotlin/com/teamyg/parfait/core/designsystem/utils/preview/PreviewComponent.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.core.ui.preview
+package com.teamyg.parfait.core.designsystem.utils.preview
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
