+++ core/designsystem/src/main/kotlin/com/teamyg/parfait/core/designsystem/utils/preview/YGPreview.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.core.ui.preview
+package com.teamyg.parfait.core.designsystem.utils.preview
 
 import androidx.compose.ui.tooling.preview.AndroidUiModes
 import androidx.compose.ui.tooling.preview.Preview
