+++ core/ui/build.gradle.kts
@@ -11,5 +11,4 @@ android {
 dependencies {
     implementation(projects.core.util.android)
     implementation(projects.core.util.jvm)
-    implementation(projects.core.designsystem)
 }
