+++ feature/camera/impl/src/main/java/com/teamyg/parfait/feature/camera/impl/component/CameraControlComponent.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.unit.dp
 import com.teamyg.parfait.feature.camera.impl.component.controls.FlipCameraButton
 import com.teamyg.parfait.feature.camera.impl.component.controls.ShutterButton
 import com.teamyg.parfait.feature.camera.impl.component.controls.ZoomLevelRow
-import com.teamyg.parfait.core.ui.preview.PreviewBox
-import com.teamyg.parfait.core.ui.preview.YGPreview
+import com.teamyg.parfait.core.designsystem.utils.preview.PreviewBox
+import com.teamyg.parfait.core.designsystem.utils.preview.YGPreview
 
 @Composable
 internal fun CameraControlComponent(
