+++ feature/camera/impl/src/main/java/com/teamyg/parfait/feature/camera/impl/component/CameraZoomIndicatorComponent.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.platform.LocalLocale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.teamyg.parfait.core.ui.preview.PreviewBox
-import com.teamyg.parfait.core.ui.preview.YGPreview
+import com.teamyg.parfait.core.designsystem.utils.preview.PreviewBox
+import com.teamyg.parfait.core.designsystem.utils.preview.YGPreview
 
 private val AccentZoomColor = Color(0xFFFFC107)
