+++ feature/gallery/impl/src/main/java/com/teamyg/parfait/feature/gallery/impl/screen/SystemGalleryPickerScreen.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import com.teamyg.parfait.feature.gallery.impl.viewmodel.SystemGalleryState
-import com.teamyg.parfait.core.ui.preview.PreviewBox
-import com.teamyg.parfait.core.ui.preview.YGPreview
+import com.teamyg.parfait.core.designsystem.utils.preview.PreviewBox
+import com.teamyg.parfait.core.designsystem.utils.preview.YGPreview
 
 @Composable
 internal fun SystemGalleryPickerScreen(
