+++ feature/login/impl/src/main/java/com/teamyg/parfait/feature/login/impl/screen/LoginScreen.kt
@@ -11,8 +11,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.teamyg.parfait.core.ui.preview.PreviewBox
-import com.teamyg.parfait.core.ui.preview.YGPreview
+import com.teamyg.parfait.core.designsystem.utils.preview.PreviewBox
+import com.teamyg.parfait.core.designsystem.utils.preview.YGPreview
 import com.teamyg.parfait.feature.login.impl.component.KakaoSignInButton
 import com.teamyg.parfait.feature.login.impl.component.OnboardingPager
 import com.teamyg.parfait.feature.login.impl.model.OnboardingPage
